Split out authentication from main
