+++ member_database/authentication/login.py
@@ -1,25 +1,15 @@
+from flask import flash, abort, request, jsonify
 from flask_login import LoginManager, login_required, current_user, AnonymousUserMixin
-from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField
-from wtforms.validators import DataRequired, EqualTo
-from flask import (
-    request, flash, abort, jsonify, current_app,
-    render_template
-)
-from itsdangerous import URLSafeTimedSerializer
 import base64
+
 from functools import wraps
 
-from .models import User
-from .mail import send_email
-from .utils import ext_url_for
-from .queries import get_user_by_name_or_email
+from .models import User, get_user_by_name_or_email
 
 
 login = LoginManager()
 
 ACCESS_LEVELS = set()
-
 
 class AnonymousUser(AnonymousUserMixin):
     '''
@@ -28,9 +18,6 @@ class AnonymousUser(AnonymousUserMixin):
     @staticmethod
     def has_access(access):
         return False
-
-
-login.anonymous_user = AnonymousUser
 
 
 def access_required(name):
@@ -101,45 +88,3 @@ def handle_needs_login():
 
     flash("You have to be logged in to access this page.", category='danger')
     abort(401)
-
-
-class LoginForm(FlaskForm):
-    user_or_email = StringField('Username/Email', validators=[DataRequired()])
-    password = PasswordField('Passwort', validators=[DataRequired()])
-    submit = SubmitField('Login')
-
-
-class SendPasswordResetForm(FlaskForm):
-    user_or_email = StringField('Username/Email', validators=[DataRequired()])
-    submit = SubmitField('Send password reset email')
-
-
-class PasswordResetForm(FlaskForm):
-    new_password = PasswordField(
-        'New password',
-        validators=[
-            DataRequired(),
-            EqualTo('confirm', message='Passwords must match')
-        ]
-    )
-    confirm = PasswordField('Confirm password', validators=[DataRequired()])
-    submit = SubmitField('Reset password')
-
-
-def send_password_reset_mail(person):
-    ts = URLSafeTimedSerializer(current_app.config["SECRET_KEY"])
-    token = ts.dumps(person.email, salt='password-reset')
-
-    link = ext_url_for('main.reset_password', token=token)
-
-    send_email(
-        subject='Password reset for registration.pep-dortmund.org',
-        recipients=[f'{person.name} <{person.email}>'],
-        sender=current_app.config['MAIL_SENDER'],
-        body=render_template('mail/reset_password.html', reset_link=link)
-    )
-
-
-def load_reset_token(token):
-    ts = URLSafeTimedSerializer(current_app.config['SECRET_KEY'])
-    return ts.loads(token, salt='password-reset', max_age=10 * 60)
