+++ member_database/models/__init__.py
@@ -1,7 +1,7 @@
 from .base import db, as_dict
-from .person import Person, User, AccessLevel, Role
+from .person import Person, MembershipStatus
 
 __all__ = [
     'db', 'as_dict',
-    'Person', 'User', 'AccessLevel', 'Role'
+    'Person', 'MembershipStatus'
 ]
